Adjust how we get/update DT
* Adjusted delta time to use a double
* Moved the delta time update to the SceneDisplay.
* Added a sleep_for to clamp to a 60 FPS frame rate.

+++ SCION_CORE/include/Core/CoreUtilities/CoreEngineData.h
@@ -8,7 +8,7 @@ namespace SCION_CORE
 class CoreEngineData
 {
   private:
-	float m_DeltaTime;
+	double m_DeltaTime;
 	float m_ScaledWidth;
 	float m_ScaledHeight;
 	float m_Gravity;
@@ -38,7 +38,7 @@ class CoreEngineData
 	const float MetersToPixels() const;
 	const float PixelsToMeters() const;
 
-	inline float GetDeltaTime() const { return m_DeltaTime; }
+	inline double GetDeltaTime() const { return m_DeltaTime; }
 	inline int WindowWidth() const { return m_WindowWidth; }
 	inline int WindowHeight() const { return m_WindowHeight; }
